+++ vice/src/piacore.c
@@ -39,12 +39,12 @@ void mypia_reset(void)
 {
    /* clear _all_ internal registers */
 
-   mypia.ctrl_a = 0;	/* PIA 1 Port A Control */
-   mypia.ctrl_b = 0;	/* PIA 1 Port B Control */
-   mypia.ddr_a = 0;	/* PIA 1 Port A DDR */
-   mypia.ddr_b = 0;	/* PIA 1 Port B DDR */
-   mypia.port_a = 255;	/* PIA 1 Port A input; nothing to read from keyboard */
-   mypia.port_b = 255;	/* PIA 1 Port B input; nothing to read from keyboard */
+   mypia.ctrl_a = 0;    /* PIA 1 Port A Control */
+   mypia.ctrl_b = 0;    /* PIA 1 Port B Control */
+   mypia.ddr_a = 0;     /* PIA 1 Port A DDR */
+   mypia.ddr_b = 0;     /* PIA 1 Port B DDR */
+   mypia.port_a = 255;  /* PIA 1 Port A input; nothing to read from keyboard */
+   mypia.port_b = 255;  /* PIA 1 Port B input; nothing to read from keyboard */
 
    pia_reset();
 
@@ -59,12 +59,12 @@ void mypia_reset(void)
 }
 
 static void mypia_update_irq(void) {
-    if( 0 
-	    || ((mypia.ctrl_a & 0x81) == 0x81)
-	    || ((mypia.ctrl_a & 0x68) == 0x48)
-	    || ((mypia.ctrl_b & 0x81) == 0x81)
-	    || ((mypia.ctrl_b & 0x68) == 0x48)
-	) {
+    if (0
+        || ((mypia.ctrl_a & 0x81) == 0x81)
+        || ((mypia.ctrl_a & 0x68) == 0x48)
+        || ((mypia.ctrl_b & 0x81) == 0x81)
+        || ((mypia.ctrl_b & 0x68) == 0x48)
+       ) {
         my_set_int(1);
     } else {
         my_set_int(0);
@@ -78,25 +78,25 @@ static void mypia_update_irq(void) {
 
 void mypia_signal(int line, int edge) {
     switch(line) {
-    case PIA_SIG_CA1:
-	if ( ((mypia.ctrl_a & 0x02) ? PIA_SIG_RISE : PIA_SIG_FALL) == edge) {
-	    mypia.ctrl_a |= 0x80;
-	    mypia_update_irq();
-	    if (IS_CA2_TOGGLE_MODE()) {
-		pia_set_ca2(1);
-		mypia.ca_state = 1;
-	    }
-	}
-    case PIA_SIG_CB1:
-	if ( ((mypia.ctrl_b & 0x02) ? PIA_SIG_RISE : PIA_SIG_FALL) == edge) {
-	    mypia.ctrl_b |= 0x80;
-	    mypia_update_irq();
-	    if (IS_CB2_TOGGLE_MODE()) {
-		pia_set_cb2(1);
-		mypia.cb_state = 1;
-	    }
-	}
-	break;
+      case PIA_SIG_CA1:
+        if (((mypia.ctrl_a & 0x02) ? PIA_SIG_RISE : PIA_SIG_FALL) == edge) {
+            mypia.ctrl_a |= 0x80;
+            mypia_update_irq();
+            if (IS_CA2_TOGGLE_MODE()) {
+                pia_set_ca2(1);
+                mypia.ca_state = 1;
+            }
+        }
+      case PIA_SIG_CB1:
+        if (((mypia.ctrl_b & 0x02) ? PIA_SIG_RISE : PIA_SIG_FALL) == edge) {
+            mypia.ctrl_b |= 0x80;
+            mypia_update_irq();
+            if (IS_CB2_TOGGLE_MODE()) {
+                pia_set_cb2(1);
+                mypia.cb_state = 1;
+            }
+        }
+        break;
     }
 }
 
@@ -117,80 +117,79 @@ void REGPARM2 mypia_store(ADDRESS addr, BYTE byte)
     addr &= 3;
 
     switch (addr) {
-
       case P_PORT_A: /* port A */
-	if (mypia.ctrl_a & 4) {
-	    mypia.port_a = byte;
-	} else {
-	    mypia.ddr_a = byte;
-	}
+        if (mypia.ctrl_a & 4) {
+            mypia.port_a = byte;
+        } else {
+            mypia.ddr_a = byte;
+        }
         byte = mypia.port_a | ~mypia.ddr_a;
-	store_pa(byte);
-	break;
+        store_pa(byte);
+        break;
 
       case P_PORT_B: /* port B */
-	if (mypia.ctrl_b & 4) {
-	    mypia.port_b = byte;
-	} else {
-	    mypia.ddr_b = byte;
-	}
+        if (mypia.ctrl_b & 4) {
+            mypia.port_b = byte;
+        } else {
+            mypia.ddr_b = byte;
+        }
         byte = mypia.port_b | ~mypia.ddr_b;
-	store_pb(byte);
-	if (IS_CB2_HANDSHAKE()) {
-	    pia_set_cb2(0);
-	    mypia.cb_state = 0;
-	    if (IS_CB2_PULSE_MODE()) {
-	        pia_set_cb2(1);
-	        mypia.cb_state = 1;
-	    }
-	}
-	break;
+        store_pb(byte);
+        if (IS_CB2_HANDSHAKE()) {
+            pia_set_cb2(0);
+            mypia.cb_state = 0;
+            if (IS_CB2_PULSE_MODE()) {
+                pia_set_cb2(1);
+                mypia.cb_state = 1;
+            }
+        }
+        break;
 
-	/* Control */
+        /* Control */
 
       case P_CTRL_A: /* Control A */
-	if ((byte & 0x38) == 0x30 ) {	/* set output low */
-	    pia_set_ca2(0);
-	    mypia.ca_state = 0;
-	} else 
-	if ((byte & 0x38) == 0x38) {	/* set output high */
-	    pia_set_ca2(1);
-	    mypia.ca_state = 1;
-	} else 				/* change to toggle/pulse */
-	if ((mypia.ctrl_a & 0x30) == 0x30) {
-	    pia_set_ca2(1);
-	    mypia.ca_state = 1;
-	} 
+        if ((byte & 0x38) == 0x30 ) {   /* set output low */
+            pia_set_ca2(0);
+            mypia.ca_state = 0;
+        } else
+        if ((byte & 0x38) == 0x38) {    /* set output high */
+            pia_set_ca2(1);
+            mypia.ca_state = 1;
+        } else                          /* change to toggle/pulse */
+        if ((mypia.ctrl_a & 0x30) == 0x30) {
+            pia_set_ca2(1);
+            mypia.ca_state = 1;
+        }
 
-	mypia.ctrl_a = (mypia.ctrl_a & 0xc0) | (byte & 0x3f);
+        mypia.ctrl_a = (mypia.ctrl_a & 0xc0) | (byte & 0x3f);
 
-	if(mypia.ctrl_a & 0x20) mypia.ctrl_a &= 0xbf;
+        if(mypia.ctrl_a & 0x20) mypia.ctrl_a &= 0xbf;
 
-	mypia_update_irq();
+        mypia_update_irq();
 
-	break;
+        break;
 
       case P_CTRL_B: /* Control B */
-	if ((byte & 0x38) == 0x30 ) {	/* set output low */
-	    pia_set_cb2(0);
-	    mypia.cb_state = 0;
-	} else 
-	if ((byte & 0x38) == 0x38) {	/* set output high */
-	    pia_set_cb2(1);
-	    mypia.cb_state = 1;
-	} else 				/* change to toggle/pulse */
-	if ((mypia.ctrl_b & 0x30) == 0x30) {
-	    pia_set_cb2(1);
-	    mypia.cb_state = 1;
-	} 
+        if ((byte & 0x38) == 0x30 ) {   /* set output low */
+            pia_set_cb2(0);
+            mypia.cb_state = 0;
+        } else
+        if ((byte & 0x38) == 0x38) {    /* set output high */
+            pia_set_cb2(1);
+            mypia.cb_state = 1;
+        } else                          /* change to toggle/pulse */
+        if ((mypia.ctrl_b & 0x30) == 0x30) {
+            pia_set_cb2(1);
+            mypia.cb_state = 1;
+        }
 
-	mypia.ctrl_b = (mypia.ctrl_b & 0xc0) | (byte & 0x3f);
+        mypia.ctrl_b = (mypia.ctrl_b & 0xc0) | (byte & 0x3f);
 
-	if(mypia.ctrl_b & 0x20) mypia.ctrl_b &= 0xbf;
+        if(mypia.ctrl_b & 0x20) mypia.ctrl_b &= 0xbf;
 
-	mypia_update_irq();
+        mypia_update_irq();
 
-	break;
+        break;
     }  /* switch */
 }
 
@@ -204,56 +203,55 @@ BYTE REGPARM1 mypia_read(ADDRESS addr)
     addr &= 3;
 
     switch (addr) {
-
       case P_PORT_A: /* port A */
-	if (mypia.ctrl_a & 4) {
+        if (mypia.ctrl_a & 4) {
 
-	    if(!is_peek_access) {
-	        mypia.ctrl_a &= 0x3f;		/* Clear CA1,CA2 IRQ */
-	        mypia_update_irq();
-	    }
-	    /* WARNING: this pin reads the voltage of the output pins, not 
-	       the ORA value as the other port. Value read might be different
-	       from what is expected due to excessive electrical load on the 
-	       pin. */
-	    byte = read_pa();
-	    pia_last_read = byte;
-	    return byte;
-	}
-	pia_last_read = (mypia.ddr_a);
-	return pia_last_read;
-	break;
+            if(!is_peek_access) {
+                mypia.ctrl_a &= 0x3f;           /* Clear CA1,CA2 IRQ */
+                mypia_update_irq();
+            }
+            /* WARNING: this pin reads the voltage of the output pins, not
+               the ORA value as the other port. Value read might be different
+               from what is expected due to excessive electrical load on the
+               pin. */
+            byte = read_pa();
+            pia_last_read = byte;
+            return byte;
+        }
+        pia_last_read = (mypia.ddr_a);
+        return pia_last_read;
+        break;
 
       case P_PORT_B: /* port B */
-	if (mypia.ctrl_b & 4) {
+        if (mypia.ctrl_b & 4) {
 
-	    if(!is_peek_access) {
-	        mypia.ctrl_b &= 0x3f;		/* Clear CB1,CB2 IRQ */
-	        mypia_update_irq();
-	    }
+            if(!is_peek_access) {
+                mypia.ctrl_b &= 0x3f;           /* Clear CB1,CB2 IRQ */
+                mypia_update_irq();
+            }
 
-	    /* WARNING: this pin reads the ORA for output pins, not 
-	       the voltage on the pins as the other port. */
-	    byte = read_pb();
-	    pia_last_read = (byte & ~mypia.ddr_b) 
-			| (mypia.port_b & mypia.ddr_b);
-	    return pia_last_read;
-	}
-	pia_last_read = (mypia.ddr_b);
-	return pia_last_read;
-	break;
+            /* WARNING: this pin reads the ORA for output pins, not
+               the voltage on the pins as the other port. */
+            byte = read_pb();
+            pia_last_read = (byte & ~mypia.ddr_b)
+                        | (mypia.port_b & mypia.ddr_b);
+            return pia_last_read;
+        }
+        pia_last_read = (mypia.ddr_b);
+        return pia_last_read;
+        break;
 
-	/* Control */
+        /* Control */
 
       case P_CTRL_A: /* Control A */
-	pia_last_read = (mypia.ctrl_a);
-	return pia_last_read;
-	break;
+        pia_last_read = (mypia.ctrl_a);
+        return pia_last_read;
+        break;
 
       case P_CTRL_B: /* Control B */
-	pia_last_read = (mypia.ctrl_b);
-	return pia_last_read;
-	break;
+        pia_last_read = (mypia.ctrl_b);
+        return pia_last_read;
+        break;
 
     }  /* switch */
 
@@ -290,11 +288,11 @@ static char snap_module_name[] = MYPIA_NAME;
  *
  * UBYTE        ORA
  * UBYTE        DDRA
- * UBYTE	CTRLA
+ * UBYTE        CTRLA
  * UBYTE        ORB
- * UBYTE	DDRB
- * UBYTE	CTRLB
- * UBYTE	CABSTATE	Bit 7 = state of CA2, Bit 6 = state of CB2
+ * UBYTE        DDRB
+ * UBYTE        CTRLB
+ * UBYTE        CABSTATE        Bit 7 = state of CA2, Bit 6 = state of CB2
  *
  */
 
@@ -317,8 +315,8 @@ int mypia_write_snapshot_module(snapshot_t * p)
     snapshot_module_write_byte(m, mypia.ddr_b);
     snapshot_module_write_byte(m, mypia.ctrl_b);
 
-    snapshot_module_write_byte(m, (mypia.ca_state ? 0x80 : 0)
-					| (mypia.cb_state ? 0x40 : 0) );
+    snapshot_module_write_byte(m, (BYTE)((mypia.ca_state ? 0x80 : 0)
+                                        | (mypia.cb_state ? 0x40 : 0)));
 
     snapshot_module_close(m);
 
@@ -331,7 +329,7 @@ int mypia_read_snapshot_module(snapshot_t * p)
     BYTE byte;
     snapshot_module_t *m;
 
-    my_restore_int(0);		/* just in case */
+    my_restore_int(0);          /* just in case */
 
     m = snapshot_module_open(p, snap_module_name, &vmajor, &vminor);
     if (m == NULL)
@@ -363,12 +361,12 @@ int mypia_read_snapshot_module(snapshot_t * p)
     byte = mypia.port_b | ~mypia.ddr_b;
     undump_pb(byte);
 
-    if( 0
-            || ((mypia.ctrl_a & 0x81) == 0x81)
-            || ((mypia.ctrl_a & 0x68) == 0x48)
-            || ((mypia.ctrl_b & 0x81) == 0x81)
-            || ((mypia.ctrl_b & 0x68) == 0x48)
-        ) {
+    if (0
+        || ((mypia.ctrl_a & 0x81) == 0x81)
+        || ((mypia.ctrl_a & 0x68) == 0x48)
+         || ((mypia.ctrl_b & 0x81) == 0x81)
+        || ((mypia.ctrl_b & 0x68) == 0x48)
+       ) {
         my_restore_int(1);
     }
